# collection: work around excessive copying of kernel BTF