… testdata/manyprogs-eb.elf
… testdata/manyprogs-el.elf