CollectionLoader: get kernel spec one time only

The internal function kernelSpec() is changed to return the kernel spec directly without making a copy. Callers of kernelSpec() such as LoadKernelSpec() are expected to make the copy. Since kernelSpec() is not public API, this does not change the public API.

This allows CORERelocate() to use the kernel spec without making an expensive copy. CORERelocate() does not patch the kernel spec so this is safe.

This PR includes a new benchmark: BenchmarkNewCollectionManyProgs. The benchmarks are done with  the following commands:
```
$ go test -exec sudo -bench=BenchmarkNewCollectionManyProgs -count 8 > master.bench
$ go test -exec sudo -bench=BenchmarkNewCollectionManyProgs -count 8 > patched.bench
$ go install golang.org/x/perf/cmd/benchstat@latest
$ benchstat master.bench patched.bench
```
Results:
```
goos: linux
goarch: amd64
pkg: github.com/cilium/ebpf
cpu: Intel(R) Core(TM) i7-6500U CPU @ 2.50GHz
                         │ master.bench │           patched.bench            │
                         │    sec/op    │   sec/op     vs base               │
NewCollectionManyProgs-4   4590.1m ± 2%   427.9m ± 4%  -90.68% (p=0.000 n=8)

                         │ master.bench  │            patched.bench            │
                         │     B/op      │     B/op      vs base               │
NewCollectionManyProgs-4   1735.8Mi ± 0%   201.6Mi ± 0%  -88.38% (p=0.000 n=8)

                         │ master.bench  │           patched.bench            │
                         │   allocs/op   │  allocs/op   vs base               │
NewCollectionManyProgs-4   11276.7k ± 0%   941.8k ± 0%  -91.65% (p=0.000 n=8)
```

cc @lmb @mauriciovasquezbernal 

Related: https://github.com/cilium/ebpf/pull/826
Partially fixes: https://github.com/cilium/ebpf/issues/917