@alban 2023-01-26
There are failures on kernel 5.10:
```
FAIL: btf TestRoundtripVMlinux
```

Is my assumption correct that CORERelocate() does not patch the kernel spec?